Improving specs to make some modules and do lint cleanup
