+++ app/assets/javascripts/components/students_table.jsx
@@ -125,4 +125,4 @@
 
   root.StudentsTable = StudentsTable;
 
-})(window)
+})(window);
