+++ app/assets/javascripts/student_profile/profile_chart_settings.jsx
@@ -43,7 +43,7 @@
         fillOpacity: 0
       }
     }
-  }
+  };
 
   ProfileChartSettings.star_chart_base_options = {
     chart: {
@@ -76,16 +76,16 @@
       dateTimeLabelFormats: {}
     },
     tooltip: {
-       formatter: function () {
-        var date = Highcharts.dateFormat('%A, %b %e, %Y', new Date(this.x))
-        var percentileRank = this.y
-        var gradeLevelEquivalent = this.points[0].point.gradeLevelEquivalent
+      formatter: function () {
+        var date = Highcharts.dateFormat('%A, %b %e, %Y', new Date(this.x));
+        var percentileRank = this.y;
+        var gradeLevelEquivalent = this.points[0].point.gradeLevelEquivalent;
 
         if ( gradeLevelEquivalent === undefined ) {
-          return date + '<br>Percentile Rank:<b> ' + percentileRank
+          return date + '<br>Percentile Rank:<b> ' + percentileRank;
         } else {
           return date + '<br>Percentile Rank:<b> ' + percentileRank +
-                 '</b><br>Grade Level Equivalent: <b>' + gradeLevelEquivalent
+                 '</b><br>Grade Level Equivalent: <b>' + gradeLevelEquivalent;
         }
       },
       shared: true
