+++ scripts/codemods/import.sh
@@ -1,0 +1,1 @@
+sed -i '' '/const dom = window.shared.ReactHelpers.dom;/d' "$TARGET"
