+++ spec/javascripts/fixtures/database_constants.jsx
@@ -1,0 +1,3 @@
+// static based on database values
+export const serviceTypesIndex = {"502":{"id":502,"name":"Attendance Officer"},"503":{"id":503,"name":"Attendance Contract"},"504":{"id":504,"name":"Behavior Contract"},"505":{"id":505,"name":"Counseling, in-house"},"506":{"id":506,"name":"Counseling, outside"},"507":{"id":507,"name":"Reading intervention"},"508":{"id":508,"name":"Math intervention"}};
+export const eventNoteTypesIndex = {"300":{"id":300,"name":"SST Meeting","created_at":"2016-03-18T00:33:49.036Z","updated_at":"2016-03-18T00:33:49.036Z"},"301":{"id":301,"name":"MTSS Meeting","created_at":"2016-03-18T00:33:49.037Z","updated_at":"2016-03-18T00:33:49.037Z"},"302":{"id":302,"name":"Parent conversation","created_at":"2016-03-18T00:33:49.038Z","updated_at":"2016-03-18T00:33:49.038Z"},"304":{"id":304,"name":"Something else","created_at":"2016-03-18T00:33:49.039Z","updated_at":"2016-03-18T00:33:49.039Z"}};
