--- spec/javascripts/fixtures/fixture_students.js
@@ -1,292 +1,0 @@
-(function() {
-  window.shared || (window.shared = {});
-  var FixtureStudents = window.shared.FixtureStudents = [{
-    "id": 1,
-    "grade": "5",
-    "hispanic_latino": true,
-    "race": "H",
-    "free_reduced_lunch": "Not Eligible",
-    "created_at": "2016-03-18T00:33:49.382Z",
-    "updated_at": "2016-03-18T00:34:20.296Z",
-    "homeroom_id": 1,
-    "first_name": "Winnie",
-    "last_name": "Mouse",
-    "state_id": null,
-    "home_language": "English",
-    "school_id": 1,
-    "student_address": null,
-    "registration_date": null,
-    "local_id": "000989",
-    "program_assigned": "Reg Ed",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "Not 504",
-    "limited_english_proficiency": "FLEP-Transitioning",
-    "most_recent_mcas_math_growth": 0,
-    "most_recent_mcas_ela_growth": 86,
-    "most_recent_mcas_math_performance": null,
-    "most_recent_mcas_ela_performance": null,
-    "most_recent_mcas_math_scaled": 233,
-    "most_recent_mcas_ela_scaled": 249,
-    "most_recent_star_reading_percentile": 74,
-    "most_recent_star_math_percentile": 90,
-    "student_risk_level": {
-      "id": 1,
-      "student_id": 1,
-      "level": 0,
-      "created_at": "2016-03-18T00:34:19.453Z",
-      "updated_at": "2016-03-18T00:34:19.475Z",
-      "mcas_math_risk_level": null,
-      "star_math_risk_level": 0,
-      "mcas_ela_risk_level": null,
-      "star_reading_risk_level": 1,
-      "limited_english_proficiency_risk_level": null
-    },
-    "discipline_incidents_count": 0,
-    "absences_count": 18,
-    "tardies_count": 22,
-    "homeroom_name": "101",
-    "event_notes": [
-      {
-        "id": 1,
-        "student_id": 1,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Lila will look into a referral to a big sister program.",
-        "recorded_at": "2010-10-21T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.596Z",
-        "updated_at": "2016-03-18T00:33:49.596Z"
-      },
-      {
-        "id": 2,
-        "student_id": 1,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Winnie twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2010-11-24T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.601Z",
-        "updated_at": "2016-03-18T00:33:49.601Z"
-      },
-      {
-        "id": 3,
-        "student_id": 1,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2011-01-23T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.604Z",
-        "updated_at": "2016-03-18T00:33:49.604Z"
-      },
-      {
-        "id": 4,
-        "student_id": 1,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Winnie will go through the special ed process for placement",
-        "recorded_at": "2011-03-24T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.607Z",
-        "updated_at": "2016-03-18T00:33:49.607Z"
-      },
-      {
-        "id": 5,
-        "student_id": 1,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2011-05-22T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.610Z",
-        "updated_at": "2016-03-18T00:33:49.610Z"
-      },
-      {
-        "id": 6,
-        "student_id": 1,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2011-07-15T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.613Z",
-        "updated_at": "2016-03-18T00:33:49.613Z"
-      }
-    ],
-    "active_services": [
-      {
-        "id": 1,
-        "student_id": 1,
-        "provided_by_educator_id": 1,
-        "recorded_by_educator_id": 1,
-        "service_type_id": 505,
-        "recorded_at": "2010-10-13T00:00:00.000Z",
-        "date_started": "2010-10-12T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.627Z",
-        "updated_at": "2016-03-18T00:33:49.627Z"
-      },
-      {
-        "id": 2,
-        "student_id": 1,
-        "provided_by_educator_id": 1,
-        "recorded_by_educator_id": 1,
-        "service_type_id": 503,
-        "recorded_at": "2010-12-07T00:00:00.000Z",
-        "date_started": "2010-12-01T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.631Z",
-        "updated_at": "2016-03-18T00:33:49.631Z"
-      },
-      {
-        "id": 3,
-        "student_id": 1,
-        "provided_by_educator_id": 1,
-        "recorded_by_educator_id": 1,
-        "service_type_id": 507,
-        "recorded_at": "2011-01-13T00:00:00.000Z",
-        "date_started": "2011-01-12T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:49.634Z",
-        "updated_at": "2016-03-18T00:33:49.634Z"
-      }
-    ],
-    "interventions": []
-  },
-  {
-    "id": 2,
-    "grade": "5",
-    "hispanic_latino": false,
-    "race": "B",
-    "free_reduced_lunch": "Not Eligible",
-    "created_at": "2016-03-18T00:33:50.423Z",
-    "updated_at": "2016-03-18T00:34:20.309Z",
-    "homeroom_id": 1,
-    "first_name": "Minnie",
-    "last_name": "Poppins",
-    "state_id": null,
-    "home_language": "Portuguese",
-    "school_id": 1,
-    "student_address": null,
-    "registration_date": null,
-    "local_id": "000722",
-    "program_assigned": "2Way English",
-    "sped_placement": null,
-    "disability": null,
-    "sped_level_of_need": null,
-    "plan_504": "Not 504",
-    "limited_english_proficiency": "FLEP-Transitioning",
-    "most_recent_mcas_math_growth": 87,
-    "most_recent_mcas_ela_growth": 52,
-    "most_recent_mcas_math_performance": "W",
-    "most_recent_mcas_ela_performance": "NI",
-    "most_recent_mcas_math_scaled": 223,
-    "most_recent_mcas_ela_scaled": 231,
-    "most_recent_star_reading_percentile": 100,
-    "most_recent_star_math_percentile": 32,
-    "student_risk_level": {
-      "id": 2,
-      "student_id": 2,
-      "level": 3,
-      "created_at": "2016-03-18T00:34:19.479Z",
-      "updated_at": "2016-03-18T00:34:19.490Z",
-      "mcas_math_risk_level": 3,
-      "star_math_risk_level": 1,
-      "mcas_ela_risk_level": 2,
-      "star_reading_risk_level": 0,
-      "limited_english_proficiency_risk_level": null
-    },
-    "discipline_incidents_count": 0,
-    "absences_count": 10,
-    "tardies_count": 15,
-    "homeroom_name": "101",
-    "event_notes": [
-      {
-        "id": 7,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2010-10-23T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.519Z",
-        "updated_at": "2016-03-18T00:33:50.519Z"
-      },
-      {
-        "id": 8,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2010-12-07T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.522Z",
-        "updated_at": "2016-03-18T00:33:50.522Z"
-      },
-      {
-        "id": 9,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 300,
-        "text": "Engagement issues in ELA, sometimes shutting down altogether and refusing to engage in academic work.  Other times this is more disruptive, lashing out and yelling at other students and interfering with their learning.",
-        "recorded_at": "2011-02-04T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.525Z",
-        "updated_at": "2016-03-18T00:33:50.525Z"
-      },
-      {
-        "id": 10,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Meredith is setting up a family meeting to discuss absences and tardies, and to follow up on concerns about supporting morning routines at home.",
-        "recorded_at": "2011-03-31T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.528Z",
-        "updated_at": "2016-03-18T00:33:50.528Z"
-      },
-      {
-        "id": 11,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Minnie twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2011-05-18T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.532Z",
-        "updated_at": "2016-03-18T00:33:50.532Z"
-      },
-      {
-        "id": 12,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 302,
-        "text": "Meetings with staff about lack of academic progress.",
-        "recorded_at": "2011-07-16T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.535Z",
-        "updated_at": "2016-03-18T00:33:50.535Z"
-      },
-      {
-        "id": 13,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Lila will look into a referral to a big sister program.",
-        "recorded_at": "2011-08-24T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.539Z",
-        "updated_at": "2016-03-18T00:33:50.539Z"
-      },
-      {
-        "id": 14,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 301,
-        "text": "Not engaged in academic work and misbehaving in class.  There have been meetings between the teacher and support staff, and Sam has talked with Minnie twice this week.  Philis following up with outside providers as well.",
-        "recorded_at": "2011-10-03T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.542Z",
-        "updated_at": "2016-03-18T00:33:50.542Z"
-      },
-      {
-        "id": 15,
-        "student_id": 2,
-        "educator_id": 1,
-        "event_note_type_id": 304,
-        "text": "Philis is meeting with parent next week and will present an attendance contract.",
-        "recorded_at": "2011-11-05T00:00:00.000Z",
-        "created_at": "2016-03-18T00:33:50.546Z",
-        "updated_at": "2016-03-18T00:33:50.546Z"
-      }
-    ],
-    "active_services": [],
-    "interventions": []
-  }];
-
-})();
