+++ spec/javascripts/student_profile/note_card_spec.jsx
@@ -1,33 +1,32 @@
-//= require ./fixtures
+import {studentProfile} from './fixtures.jsx';
+import SpecSugar from '../support/spec_sugar.jsx';
 
 describe('NoteCard', function() {
-  var dom = window.shared.ReactHelpers.dom;
-  var createEl = window.shared.ReactHelpers.createEl;
-  var merge = window.shared.ReactHelpers.merge;
+  const merge = window.shared.ReactHelpers.merge;
+  const ReactDOM = window.ReactDOM;
+  const NoteCard = window.shared.NoteCard;
+  const moment = window.moment;
 
-  var SpecSugar = window.shared.SpecSugar;
-  var NoteCard = window.shared.NoteCard;
-  var Fixtures = window.shared.Fixtures;
-  var moment = window.moment;
-
-  var helpers = {
+  const helpers = {
     renderInto: function(el, props) {
-      var mergedProps = merge(props || {}, {
+      const mergedProps = merge(props || {}, {
         noteMoment: moment(),
         educatorId: 1,
-        badge: dom.span({}, ''),
+        badge: <span>
+          {''}
+        </span>,
         onSave: jasmine.createSpy('onSave'),
         eventNoteId: 1,
         eventNoteTypeId: 1,
-        educatorsIndex: Fixtures.studentProfile.educatorsIndex,
+        educatorsIndex: studentProfile.educatorsIndex,
         attachments: []
       });
 
-      return ReactDOM.render(createEl(NoteCard, mergedProps), el);
+      return ReactDOM.render(<NoteCard {...mergedProps} />, el); //eslint-disable-line react/no-render-return-value
     },
 
     editNoteAndSave: function(el, uiParams) {
-      var $text = $(el).find('.note-text');
+      const $text = $(el).find('.note-text');
       $text.html(uiParams.html);
       React.addons.TestUtils.Simulate.input($text.get(0));
       React.addons.TestUtils.Simulate.blur($text.get(0));
@@ -41,7 +40,7 @@ describe('NoteCard', function() {
 
   SpecSugar.withTestEl('render', function() {
     it('renders simple text', function() {
-      var el = this.testEl;
+      const el = this.testEl;
 
       helpers.renderInto(el, {
         text: 'hello'
@@ -51,7 +50,7 @@ describe('NoteCard', function() {
     });
 
     it('renders number of revisions', function() {
-      var el = this.testEl;
+      const el = this.testEl;
 
       helpers.renderInto(el, {
         text: 'hello',
@@ -62,7 +61,7 @@ describe('NoteCard', function() {
     });
 
     it('escapes HTML-meaningful characters in text', function() {
-      var el = this.testEl;
+      const el = this.testEl;
 
       helpers.renderInto(el, {
         text: 'hello <script src="xss.js"></script>world'
@@ -72,7 +71,7 @@ describe('NoteCard', function() {
     });
 
     it('renders newlines as <br> tags', function() {
-      var el = this.testEl;
+      const el = this.testEl;
 
       helpers.renderInto(el, {
         text: 'hello\nworld'
@@ -84,9 +83,9 @@ describe('NoteCard', function() {
 
   SpecSugar.withTestEl('integration tests', function() {
     it('replaces HTML with newlines in saved text', function() {
-      var el = this.testEl;
+      const el = this.testEl;
 
-      var component = helpers.renderInto(el, {
+      const component = helpers.renderInto(el, {
         text: 'hello world'
       });
 
@@ -102,9 +101,9 @@ describe('NoteCard', function() {
     });
 
     it('sanitizes undesirable HTML', function() {
-      var el = this.testEl;
+      const el = this.testEl;
 
-      var component = helpers.renderInto(el, {
+      const component = helpers.renderInto(el, {
         text: 'hello\nworld'
       });
