+++ spec/javascripts/student_profile/page_container_spec.jsx
@@ -1,15 +1,12 @@
-//= require ./fixtures
+import {nowMoment, studentProfile} from './fixtures.jsx';
+import SpecSugar from '../support/spec_sugar.jsx';
 
 describe('PageContainer', function() {
-  var dom = window.shared.ReactHelpers.dom;
-  var createEl = window.shared.ReactHelpers.createEl;
-  var merge = window.shared.ReactHelpers.merge;
+  const merge = window.shared.ReactHelpers.merge;
+  const ReactDOM = window.ReactDOM;
+  const PageContainer = window.shared.PageContainer;
 
-  var SpecSugar = window.shared.SpecSugar;
-  var PageContainer = window.shared.PageContainer;
-  var Fixtures = window.shared.Fixtures;
-
-  var helpers = {
+  const helpers = {
     findColumns: function(el) {
       return $(el).find('.summary-container > div');
     },
@@ -38,15 +35,15 @@ describe('PageContainer', function() {
     },
 
     renderInto: function(el, props) {
-      var mergedProps = merge(props || {}, {
-        nowMomentFn: function() { return Fixtures.nowMoment; },
-        serializedData: Fixtures.studentProfile,
+      const mergedProps = merge(props || {}, {
+        nowMomentFn: function() { return nowMoment; },
+        serializedData: studentProfile,
         queryParams: {},
         history: SpecSugar.history(),
         actions: helpers.createSpyActions(),
         api: helpers.createSpyApi()
       });
-      return ReactDOM.render(createEl(PageContainer, mergedProps), el);
+      ReactDOM.render(<PageContainer {...mergedProps} />, el);
     },
 
     takeNotesAndSave: function(el, uiParams) {
@@ -57,8 +54,8 @@ describe('PageContainer', function() {
     },
 
     editNoteAndSave: function(el, uiParams) {
-      var $noteCard = $(el).find('.NotesList .NoteCard').first();
-      var $text = $noteCard.find('.note-text');
+      const $noteCard = $(el).find('.NotesList .NoteCard').first();
+      const $text = $noteCard.find('.note-text');
       $text.html(uiParams.text);
       React.addons.TestUtils.Simulate.input($text.get(0));
       React.addons.TestUtils.Simulate.blur($text.get(0));
@@ -75,7 +72,7 @@ describe('PageContainer', function() {
 
   SpecSugar.withTestEl('integration tests', function() {
     it('renders everything on the happy path', function() {
-      var el = this.testEl;
+      const el = this.testEl;
       helpers.renderInto(el);
 
       expect(el).toContainText('Daisy Poppins');
@@ -83,7 +80,7 @@ describe('PageContainer', function() {
       expect($(el).find('.Sparkline').length).toEqual(9);
       expect($(el).find('.InterventionsDetails').length).toEqual(1);
 
-      var interventionLists = helpers.interventionSummaryLists(el);
+      const interventionLists = helpers.interventionSummaryLists(el);
       expect(interventionLists.length).toEqual(3);
       expect(interventionLists[0]).toContainText('Reg Ed');
       expect(interventionLists[0]).toContainText('Homeroom 102');
@@ -92,7 +89,7 @@ describe('PageContainer', function() {
     });
 
     it('opens dialog when clicking Take Notes button', function() {
-      var el = this.testEl;
+      const el = this.testEl;
       helpers.renderInto(el);
 
       $(el).find('.btn.take-notes').click();
@@ -101,7 +98,7 @@ describe('PageContainer', function() {
     });
 
     it('opens dialog when clicking Record Service button', function() {
-      var el = this.testEl;
+      const el = this.testEl;
       helpers.renderInto(el);
 
       $(el).find('.btn.record-service').click();
@@ -110,8 +107,8 @@ describe('PageContainer', function() {
     });
 
     it('can save notes for SST meetings, mocking the action handlers', function() {
-      var el = this.testEl;
-      var component = helpers.renderInto(el);
+      const el = this.testEl;
+      const component = helpers.renderInto(el);
       helpers.takeNotesAndSave(el, {
         eventNoteTypeText: 'SST Meeting',
         text: 'hello!'
@@ -125,8 +122,8 @@ describe('PageContainer', function() {
     });
 
     it('can edit notes for SST meetings, mocking the action handlers', function() {
-      var el = this.testEl;
-      var component = helpers.renderInto(el);
+      const el = this.testEl;
+      const component = helpers.renderInto(el);
 
       helpers.editNoteAndSave(el, {
         eventNoteTypeText: 'SST Meeting',
@@ -141,8 +138,8 @@ describe('PageContainer', function() {
     });
 
     it('verifies that the educator name is in the correct format', function() {
-      var el = this.testEl;
-      var component = helpers.renderInto(el, {});
+      const el = this.testEl;
+      const component = helpers.renderInto(el, {});
 
       // Simulate that the server call is still pending
       component.props.api.saveService.and.returnValue($.Deferred());
@@ -183,9 +180,9 @@ describe('PageContainer', function() {
     // });
 
     it('#mergedDiscontinueService', function() {
-      var el = this.testEl;
-      var instance = helpers.renderInto(el);
-      var updatedState = instance.mergedDiscontinueService(instance.state, 312, 'foo');
+      const el = this.testEl;
+      const instance = helpers.renderInto(el);
+      const updatedState = instance.mergedDiscontinueService(instance.state, 312, 'foo');
       expect(Object.keys(updatedState)).toEqual(Object.keys(instance.state));
       expect(updatedState.requests.discontinueService).toEqual({ 312: 'foo' });
       expect(instance.mergedDiscontinueService(updatedState, 312, null)).toEqual(instance.state);
