+++ spec/javascripts/student_profile/profile_bar_chart_spec.jsx
@@ -1,28 +1,29 @@
-describe('ProfileBarCharts', function() {
-  var ProfileBarChart = window.shared.ProfileBarChart;
-  var createEl = window.shared.ReactHelpers.createEl;
-  var merge = window.shared.ReactHelpers.merge;
-  var SpecSugar = window.shared.SpecSugar;
+import SpecSugar from '../support/spec_sugar.jsx';
 
-  var helpers = {
+describe('ProfileBarCharts', function() {
+  const ProfileBarChart = window.shared.ProfileBarChart;
+  const merge = window.shared.ReactHelpers.merge;
+  const ReactDOM = window.ReactDOM;
+  
+  const helpers = {
     renderInto: function(el, props) {
-      var mergedProps = merge({
+      const mergedProps = merge({
         events: [],
         id: 'foo-id',
         titleText: 'foo-title',
         monthsBack: 48,
         nowMomentUTC: moment.utc('2017-02-02T13:23:15+00:00')
       }, props || {});
-      return ReactDOM.render(createEl(ProfileBarChart, mergedProps), el);
+      ReactDOM.render(<ProfileBarChart {...mergedProps} />, el);
     }
   };
 
   SpecSugar.withTestEl('integration tests', function() {
     it('is wrapped in a div with the given id', function() {
-      var el = this.testEl;
+      const el = this.testEl;
       helpers.renderInto(el, {id: 'foo'});
 
-      var div = $(el).children().first();
+      const div = $(el).children().first();
       expect(div.attr('id')).toEqual("foo");
     });
