+++ spec/javascripts/student_profile/services_list_spec.jsx
@@ -1,16 +1,14 @@
-//= require ./fixtures
+import {studentProfile} from './fixtures.jsx';
+import SpecSugar from '../support/spec_sugar.jsx';
+
 
 describe('ServicesList', function() {
-  var dom = window.shared.ReactHelpers.dom;
-  var createEl = window.shared.ReactHelpers.createEl;
-  var merge = window.shared.ReactHelpers.merge;
+  const merge = window.shared.ReactHelpers.merge;
+  const ReactDOM = window.ReactDOM;
+  const ServicesList = window.shared.ServicesList;
+  const fromPair = window.shared.fromPair;
 
-  var SpecSugar = window.shared.SpecSugar;
-  var ServicesList = window.shared.ServicesList;
-  var fromPair = window.shared.fromPair;
-  var Fixtures = window.shared.Fixtures;
-
-  var helpers = {
+  const helpers = {
     emptyServicesFeed: function() {
       return { active: [], discontinued: [] };
     },
@@ -37,29 +35,29 @@ describe('ServicesList', function() {
     },
 
     renderInto: function(el, props) {
-      var mergedProps = merge({
+      const mergedProps = merge({
         servicesFeed: {
           active: [],
           discontinued: []
         },
-        educatorsIndex: Fixtures.studentProfile.educatorsIndex,
-        serviceTypesIndex: Fixtures.studentProfile.serviceTypesIndex,
+        educatorsIndex: studentProfile.educatorsIndex,
+        serviceTypesIndex: studentProfile.serviceTypesIndex,
         discontinueServiceRequests: {},
         onClickDiscontinueService: jasmine.createSpy('onClickDiscontinueService')
       }, props || {});
-      return ReactDOM.render(createEl(ServicesList, mergedProps), el);
+      ReactDOM.render(<ServicesList {...mergedProps} />, el);
     }
   };
 
   SpecSugar.withTestEl('high-level integration tests', function() {
     it('renders message when no services', function() {
-      var el = this.testEl;
+      const el = this.testEl;
       helpers.renderInto(el, { servicesFeed: helpers.emptyServicesFeed() });
       expect(el).toContainText('No services');
     });
 
     it('renders everything on the happy path', function() {
-      var el = this.testEl;
+      const el = this.testEl;
       helpers.renderInto(el, { servicesFeed: helpers.oneActiveServiceFeed() });
       expect(el).toContainText('Reading intervention');
       expect(el).toContainText('With');
@@ -68,7 +66,7 @@ describe('ServicesList', function() {
     });
 
     it('asks for confirmation before discontinuing', function() {
-      var el = this.testEl;
+      const el = this.testEl;
       helpers.renderInto(el, { servicesFeed: helpers.oneActiveServiceFeed() });
       $(el).find('.btn').click();
       expect(el).toContainText('Confirm');
@@ -76,8 +74,8 @@ describe('ServicesList', function() {
     });
 
     it('shows a message when request in progress', function() {
-      var el = this.testEl;
-      var service = helpers.fixtureService();
+      const el = this.testEl;
+      const service = helpers.fixtureService();
       helpers.renderInto(el, {
         servicesFeed: helpers.oneActiveServiceFeed(),
         discontinueServiceRequests: fromPair(service.id, 'pending')
@@ -86,8 +84,8 @@ describe('ServicesList', function() {
     });
 
     it('renders discontinued services correctly', function() {
-      var el = this.testEl;
-      var discontinuedService = merge(helpers.fixtureService(), {
+      const el = this.testEl;
+      const discontinuedService = merge(helpers.fixtureService(), {
         discontinued_by_educator_id: 1,
         discontinued_recorded_at: "2016-04-05T01:43:15.256Z"
       });
